backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
langchain
langchain-groq
python-jose
passlib[bcrypt]
sentry-sdk
pywebpush
cryptography
pypdf
faiss-cpu
langchain-community
sentence-transformers
gunicorn